requirements.txt
groq
gradio
pandas
pdfplumber
python-docx